# LSM: Approximate pacing